backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0
channels>=4.0.0
channels-redis>=4.1.0
redis>=4.6.0
celery>=5.3.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
stripe>=6.7.0
boto3>=1.28.0
django-storages>=1.14.0
python-decouple>=3.8
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0
drf-yasg>=1.21.0
factory-boy>=3.3.0
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0
flake8>=6.0.0
black>=23.9.0
isort>=5.12.0
pre-commit>=3.4.0
gunicorn>=21.2.0
whitenoise>=6.6.0
sentry-sdk>=1.32.0
django-ratelimit>=4.1.0
django-prometheus>=2.3.0